Safari-609.1.11.1/LayoutTests/imported/w3c/web-platform-tests/xhr/getresponseheader-error-state-expected.txt

Blocked access to external URL http://www1.localhost:8800/xhr/resources/nocors/folder.txt
CONSOLE MESSAGE: line 32: XMLHttpRequest cannot load http://www1.localhost:8800/xhr/resources/nocors/folder.txt due to access control checks.

PASS XMLHttpRequest: getResponseHeader() in error state (failing cross-origin test) 
